Boas práticas e outras coisas
Today i learned that i can't implement a function inside a function and call the function that is inside the fucking function.
Well... I shoud've know.

As the functions of the Car Class would always point to None, i decided to set it to point to a str instead. It works.