update: I finally get it!
So... to convert an exception into another, u need to insert both of them inside a Try/Except block, but the position of this statement is all that matters! If u want it to return a KeyboardInterrupt, u need to put the var inside a try/except block and then u return a KeyboardInterrupt error with a better SystemExit message!

I finally understand!!!