updated tryCatch.py
Added a validation While lasso loop, where it prevents the user from inserting only spaces or a blank input by using the .strip() method in the e_INPUT prompted to the user.

If we insert a if PARAMETER: followed by a break statement, it just consider there's nothing for it and break the empty if statement. As it is outside the while loop, it wont break the loop itself (yet).